refactor(ai): extract recipe scoring helpers into recipe_scoring.dart (Refs #2289)
`economy_planner.dart` previously mixed labour allocation / cargo
preference with recipe feasibility and scoring. This matches issue
#2289 proposed work item 5: move `_ScoredRecipe`, `_feasibleRuns`,
`_scoreRecipe`, `_recipeChainScore`, `_recipeAgendaScore`, and the
recipe-scoring constants into a focused `recipe_scoring.dart` sibling
under `planning/`.

Helpers used only by `economy_planner.dart` keep the same shape; the
two chain/agenda scoring functions stay file-private. No behavior
change: feasibility, scoring, allocation, and cargo preference are
deterministic for the same inputs.

Refs #2289
